pyproject.toml
[tool.poetry]
name = "replica1"
version = "0.1.0"
description = "A LLM controls application"
authors = ["Your Name <devashishthapliyal1@gmail.com>"]
packages = [
    {include = "backend", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.110.0"
uvicorn = "0.27.1"
langchain = "0.1.9"
langchain-openai = "0.0.7"
pydantic = "1.10.8"
python-dotenv = "1.0.1"
requests = "2.31.0"
numpy = "1.24.4"
pandas = "1.5.3"
websockets = "12.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
pytest-asyncio = "0.23.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 